made page so user can choose how to create their recipe (generating or creating their own)
