+++ package.json
@@ -12,7 +12,7 @@
     "@google/generative-ai": "^0.24.0",
     "axios": "^1.8.4",
     "bcryptjs": "^3.0.2",
-    "framer-motion": "^12.6.2",
+    "framer-motion": "^12.6.3",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.12.1",
     "multer": "^1.4.5-lts.2",
