+++ src/app/create_recipe/page.tsx
@@ -1,204 +1,94 @@
 "use client";
-import { useState } from "react";
-import axios from "axios";
-import { toast } from "react-hot-toast";
+
+import Link from "next/link";
 import Image from "next/image";
-import Button from "@/components/Button";
+import { motion } from "framer-motion";
 
-export default function CreateRecipe() {
-  const [title, setTitle] = useState("");
-  const [ingredients, setIngredients] = useState("");
-  const [ingredientList, setIngredientList] = useState<string[]>([]);
-  const [instructions, setInstructions] = useState("");
-  const [cookingTime, setCookingTime] = useState("");
+const cardVariants = {
+  hidden: { opacity: 0, y: 50 },
+  visible: {
+    opacity: 1,
+    y: 0,
+    transition: { duration: 0.6, ease: "easeOut" },
+  },
+};
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    // Validate inputs
-    if (
-      !title ||
-      ingredientList.length === 0 ||
-      !instructions ||
-      !cookingTime
-    ) {
-      toast.error("Please fill all fields");
-      return;
-    }
-
-    const cookingTimeNumber = parseInt(cookingTime);
-    if (isNaN(cookingTimeNumber)) {
-      toast.error("Cooking time must be a number");
-      return;
-    }
-
-    try {
-      await axios.post("/api/recipes/createRecipe", {
-        title,
-        ingredients: ingredientList,
-        instructions,
-        cookingTime: cookingTimeNumber,
-      });
-
-      toast.success("Recipe created successfully!");
-
-      // Reset form after success
-      setTitle("");
-      setIngredients("");
-      setIngredientList([]);
-      setInstructions("");
-      setCookingTime("");
-    } catch (error: any) {
-      toast.error("Error creating recipe. Please try again later.");
-    }
-  };
-
-  const handleAddIngredient = () => {
-    if (ingredients.trim() !== "") {
-      setIngredientList((prev) => [...prev, ingredients.trim()]);
-      setIngredients(""); // Curățăm câmpul de input
-    }
-  };
-
-  const handleRemoveIngredient = (index: number) => {
-    setIngredientList((prev) => prev.filter((_, i) => i !== index));
-  };
-
+const CreateRecipe = () => {
   return (
-    <div className="max-w-3xl mx-auto p-6 mt-10">
-      <div className="max-w-7xl mx-auto">
-        <div className="text-center mb-10">
-          <h1 className="text-3xl md:text-4xl font-bold text-gray-800 mb-4">
-            Recipe Oasis
-          </h1>
-          <p className="text-lg text-gray-600 max-w-2xl mx-auto">
-            Create and save your favorite recipes in one place. Share with
-            friends or keep them as your culinary secrets.
-          </p>
-        </div>
-      </div>
-      <div className="bg-gray-50 mb-10 p-8 rounded-lg shadow-xl">
-        <div className="flex items-center justify-center mb-6">
-          <Image
-            src="/chef_hat_heart.svg"
-            alt="chef-hat"
-            width={25}
-            height={25}
-          />
-          <h2 className="text-2xl font-bold text-gray-800 pl-2">
-            Create New Recipe
-          </h2>
-        </div>
-        <form onSubmit={handleSubmit}>
-          {/* Title Field */}
-          <div className="mb-5">
-            <h3 className="text-center mb-2">Recipe Title</h3>
-            <div className="relative">
+    <div className="flex flex-col items-center justify-center min-h-screen bg-gradient-to-br from-green-100 to-blue-100 px-4 py-16">
+      <motion.h1
+        initial={{ opacity: 0, y: -30 }}
+        whileInView={{ opacity: 1, y: 0 }}
+        transition={{ duration: 0.6 }}
+        viewport={{ once: true }}
+        className="text-4xl font-bold text-gray-800 text-center mb-4"
+      >
+        Your Digital Kitchen
+      </motion.h1>
+
+      <motion.p
+        initial={{ opacity: 0, y: -20 }}
+        whileInView={{ opacity: 1, y: 0 }}
+        transition={{ delay: 0.2, duration: 0.6 }}
+        viewport={{ once: true }}
+        className="text-lg text-gray-600 max-w-xl text-center mb-12"
+      >
+        Create perfect recipes with the help of artificial intelligence or add
+        and organize your own favorite recipes.
+      </motion.p>
+
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-10 w-full max-w-4xl px-4">
+        {/* AI Card */}
+        <Link href="/create_recipe/ai">
+          <motion.div
+            variants={cardVariants}
+            initial="hidden"
+            whileInView="visible"
+            viewport={{ once: true }}
+            className="rounded-2xl shadow-md bg-white p-8 flex flex-col items-center text-center hover:shadow-lg"
+          >
+            <div className="w-16 h-16 flex items-center justify-center rounded-full bg-gradient-to-br from-green-500 to-blue-500 mb-4">
+              <Image src="/wand.svg" alt="AI Chef" width={32} height={32} />
+            </div>
+            <h2 className="text-3xl font-semibold text-gray-800 mb-2">
+              Generate with AI
+            </h2>
+            <p className="text-gray-600 text-md">
+              Let artificial intelligence suggest recipes based on the
+              ingredients you have or your culinary preferences.
+            </p>
+          </motion.div>
+        </Link>
+
+        {/* Manual Card */}
+        <Link href="/create_recipe/manual">
+          <motion.div
+            variants={cardVariants}
+            initial="hidden"
+            whileInView="visible"
+            viewport={{ once: true }}
+            className="rounded-2xl shadow-md bg-white p-8 flex flex-col items-center text-center hover:shadow-lg"
+          >
+            <div className="w-16 h-16 flex items-center justify-center rounded-full bg-gradient-to-br from-green-500 to-blue-500 mb-4">
               <Image
-                src="/fork-knife.svg"
-                alt="fork-knife-icon"
-                width={20}
-                height={20}
-                className="absolute left-3 top-1/2 transform -translate-y-1/2"
-              />
-              <input
-                type="text"
-                value={title}
-                onChange={(e) => setTitle(e.target.value)}
-                placeholder="e.g. Grandma's Famous Apple Pie"
-                className="w-full p-3 pl-10 bg-gray-100 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-400"
+                src="/book.svg"
+                alt="Manual Recipe"
+                width={32}
+                height={32}
               />
             </div>
-          </div>
-
-          {/* Ingredients Field */}
-          <div className="mb-4">
-            <h3 className="text-center mb-2">Ingredients</h3>
-            <div className="flex items-center space-x-3">
-              <input
-                type="text"
-                value={ingredients}
-                onChange={(e) => setIngredients(e.target.value)}
-                placeholder="Enter an ingredient"
-                className="w-full p-3 pl-5 bg-gray-100 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-400"
-              />
-              <Button
-                type="button"
-                icon="plus.svg"
-                title="Add"
-                variant="btn_small_gradient p-3 pr-6 sm: pr-10"
-                onClick={handleAddIngredient}
-              />
-            </div>
-          </div>
-
-          {/* Display Added Ingredients */}
-          {ingredientList.length > 0 && (
-            <div className="mb-4 flex flex-wrap gap-2">
-              {ingredientList.map((ingredient, index) => (
-                <div
-                  key={index}
-                  className="flex items-center bg-gray-200 px-4 py-2 rounded-lg"
-                >
-                  <span className="text-sm">{ingredient}</span>
-                  <button
-                    type="button"
-                    onClick={() => handleRemoveIngredient(index)}
-                    className="ml-2 text-red-500 hover:text-red-700 cursor-pointer"
-                  >
-                    <Image
-                      src="/close.svg"
-                      alt="close-icon"
-                      width={16}
-                      height={16}
-                    />
-                  </button>
-                </div>
-              ))}
-            </div>
-          )}
-
-          {/* Instructions Field */}
-          <div className="mb-4">
-            <h3 className="text-center mb-2">Cooking Instructions</h3>
-            <textarea
-              value={instructions}
-              onChange={(e) => setInstructions(e.target.value)}
-              placeholder="Describe the preparation and cooking steps in detail.."
-              className="w-full p-3 bg-gray-100 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-400"
-              rows={5}
-            />
-          </div>
-
-          {/* Cooking Time Field */}
-          <div className="mb-4">
-            <h3 className="text-center mb-2">Cooking Time (minutes)</h3>
-            <div className="relative flex items-center space-x-3">
-              <Image
-                src="/clock.svg"
-                alt="clock-icon"
-                width={20}
-                height={20}
-                className="absolute left-3 top-1/2 transform -translate-y-1/2"
-              />
-              <input
-                type="number"
-                value={cookingTime}
-                onChange={(e) => setCookingTime(e.target.value)}
-                placeholder="e.g. 45"
-                className="w-full p-3 pl-10 bg-gray-100 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-400"
-              />
-            </div>
-          </div>
-
-          {/* Submit Button */}
-          <Button
-            type="submit"
-            title="Create Recipe"
-            variant="btn_small_gradient mx-auto block"
-          />
-        </form>
+            <h2 className="text-3xl font-semibold text-gray-800 mb-2">
+              Add your own recipes
+            </h2>
+            <p className="text-gray-600 text-md">
+              Prefer the classic way? Add your ingredients and instructions step
+              by step.
+            </p>
+          </motion.div>
+        </Link>
       </div>
     </div>
   );
-}
+};
+
+export default CreateRecipe;
